Move traits to a separate file.
