+++ crates/bounce/src/query/mod.rs
@@ -21,9 +21,11 @@ mod prepared_query;
 mod query_;
 mod query_value;
 mod status;
+mod traits;
 
-pub use mutation::{use_mutation_value, Mutation, MutationResult, UseMutationValueHandle};
+pub use mutation::{use_mutation_value, UseMutationValueHandle};
 pub use prepared_query::use_prepared_query;
-pub use query_::{use_query, Query, QueryResult, UseQueryHandle};
+pub use query_::{use_query, UseQueryHandle};
 pub use query_value::{use_query_value, UseQueryValueHandle};
 pub use status::QueryStatus;
+pub use traits::{Mutation, MutationResult, Query, QueryResult};
