+++ crates/bounce/src/query/prepared_query.rs
@@ -6,9 +6,10 @@ use yew::prelude::*;
 use yew::suspense::{Suspension, SuspensionResult};
 
 use super::query_::{
-    Query, QuerySelector, QueryState, QueryStateAction, QueryStateValue, RunQuery, RunQueryInput,
+    QuerySelector, QueryState, QueryStateAction, QueryStateValue, RunQuery, RunQueryInput,
     UseQueryHandle,
 };
+use super::traits::Query;
 use crate::states::future_notion::use_future_notion_runner;
 use crate::states::input_selector::use_input_selector_value;
 use crate::states::slice::use_slice_dispatch;
